Refactored chat system to use an internal callback system.
Refactored the chat system to use an internal callback system, because
I decided to switch away from SF.WrapFunction for this.

I need to know specifically what functions any given instance is listening
to the chat system on. So that I can make sure to remove references to any
functions from that instance when it deinitializes. And since the table of
wrapped functions is local to sflib, I had to refactor away from WrapFunction.