Fix duplicating
Duplicating was not working because table.Copy has a problem with
duplicating tables that have __metatable set to a non-table.
The only way I found to fix this is by replacing table.Copy altogether,
because adv. duplicator is using table.Copy and the original entity
cannot be modified for security reasons. I made a PR to garry to fix his
table.Copy, once that is accepted, that part can be removed.

In addition, screens were using outdated fields.
Fixes #56