Fix createFromImage bad path check
In material.createFromImage, Material() will never receive a valid path if used from the data/ folder. 

The current implementation prepends "data/" to the path, then checks if it exists in game using file.Exists with sf.NormalizePath 
This means that if you pass "data/sf_filedata/mymaterial.png" to the current function, the game will check to see if "data/data/sf_filedata/mymaterial.png" exists. 

The file will obviously not exist (unless it simultaneously exists in /data/data/%dest and /data/%dest)  

The function doesn't prepend /data/ to the absolute path it is passing to material. 


The proposed changes now check to see if the first five characters of the path string are "/data", then performs the check on the absolute path if that condition is true (this also saves on IO operations)