tscm_compat: Introduce "tscm"
This is a basis for future work to allow the 'TSCM' Starfall variant's scripts to be executed on StarfallEx.

It adds a preprocessor directive to enable TSCM compatibility across the script (there are some mutually incompatible parts of the TSCM implementation not shown here).

TSCM compatibility is achieved via a shared module which performs the required wrapping using other existing modules to try and avoid increasing the API's surface area.

(Theoretically, this would be doable with completely in-Starfall code, but I wasn't sure how to do that.)

This adds 4 basic aliases on the Entity type (:class(), :pos(), :ang(), :owner()),
along with all the aliases for the `ents` and `time` libraries.

The idea behind the structure is that this would severely bloat the documentation if applied directly.