Make the documentation for --@clientmain more clear
I was convinced this directive was broken for MONTHS. I was actually just about to file a bug report when I realized that I was failing to --@include the file the whole time. I don't want anyone else to be hitting their head against the wall like I was. (I verified that the main file is not sent to the client by checking instance.source and instance.scripts. Hopefully, the days of networking secrets to the chip via cslua are over.)