Destroy documentation
--@param and --@return docs have all
/ will all be changed.

They should be in the format --@param <type> <param_name> <desc> now. Same goes for --@return, but without the param name.

For now it uses the description to put the type in there, in the future, we want to edit the dochtml branch to add highlighted text for the type or something.
Still missing these:
* material library / type
* render library / type
Everything else should be done.
Socket library also got some docs.
Bit library got docs.
Everything gets docs.