Use Docs.Types instead of SF.Types
Realised that this won't run in the glua environment. I thought this might give us problems with param and return docs showing up before type docs, but it seems to not have that issue?

I'll look more into it later.

Also got rid of the use of string.Explode in the doc, that build_docs.lua in docgen creates. This shouldn't cause any problems.

Other fixes:
* I missed documentation for some quat metamethods
* Vehicle, Weapons and NPC docs were missing their metatable as a libtbl leading to errors when documenting their __tostring methods.
* Was missing thread in the generic lua type lookup table