Add 'identifier' keyword to editor themes as well as separate constant values from keywords (#1136)
* Add the 'identifier' token to the editor + constant keywords

This modifies the editor to make any text that can be a variable an 'identifier'.

Also _G, self, ..., nil, true and false are now constants.

For the default theme you'll only notice the new constant keywords being highlighted as constants instead of keywords.

* 'do' can't have parenthesis next to it

Saw it had the structure of being able to be beside parenthesis, but maybe that was a mistake.

* Add string library highlighting

Previously using string syntax sugar like
(""):rep(5) wouldn't highlight ``rep`` since strings aren't a type. This adds all the methods from the string library to properly highlight.