Include dummy argument in autocomplete
Some users requested that the command autocomplete display the name of the target user. So to do this I am introducing a 2nd argument that will do nothing for the time being.

In the future, it may be worth considering other options such as the ULX route: Name is the default argument, but prepend a `$` and it becomes a SteamID check.