Allow EntityFireBullets hook to set a bullet callback (#1553)
* EntityFireBullets hook bullet callback

Allows for the EntityFireBullets hook to optionally return a function that will be called as if it were the Bullet.Callback. When the bullet gets fired, the callback will be called with the wrapped attacker and traceResult, dmgInfo (CTakeDamageInfo) has been omitted.

* Chain callbacks

I had a hunch this could be an issue so I tested it. This commit daisy chains existing callbacks set on the Bullet structure so they don't get overwritten and everything gets called

* Run hook last, use instance.runFunction

Returns true on SF.hookAdd to run hook last, preventing messing with the hook chain and calls the callback wrapped with instance.runFunction now

* If someone set a callback they probably return true anyway and this won't run

---------

Co-authored-by: Garrett Brown <grbrown93@sbcglobal.net>