Url whitelist changes (#1760)
* Fixed Styling

Made all the spacing of the urls added and comments all the same instead of some having a space after them.

* Added URL

This url is to a Youtube to mp4 / mp3 converter api. I have used this converter a lot inside gmod and it works well. There is a free plan as well as a paid plan for people who want more request / longer songs.

* Fixed More Styling

Made all the spacing of the urls and comments all the same instead of some having a space after them.

* Fixed url whitelist

Updated example to a one of my converted video files.