Fix clientside starfall instances not cleaning up after a fullupdate
Tested on a fresh dedicated server with only the latest version of starfall installed. A starfall clientside instance fails to clean up after a fullupdate occurs. This is because the removedhook responsible for deinitialzing it is removed any time EntityRemoved is called. This is okay serverside, but clientside EntityRemoved can be called during a fullupdate. The entity still counts as valid, so the clientside instance is not removed by the fullupdate, however the removedhook is set to nil regardless, causing the clientside instance to now remain permanently.